(ui): [Settings] Started shift to new button action bar styling
