+++ lib/core/ui/deprecated/messages/generic.dart
@@ -36,7 +36,8 @@ class LSGenericMessage extends StatelessWidget {
                         LunaButton.text(
                             text: buttonText,
                             onTap: onTapHandler,
-                            backgroundColor: LunaColours.accent,
+                            backgroundColor: Theme.of(context).cardColor,
+                            color: LunaColours.accent,
                         ),
                     ],
                 ),
