+++ lib/core/utilities/changelog.dart
@@ -67,6 +67,7 @@ class LunaChangelog {
                         actions: [
                             LunaButton.text(
                                 text: 'Full Changelog',
+                                icon: Icons.track_changes_rounded,
                                 onTap: () async => Constants.URL_CHANGELOG.lunaOpenGenericLink(),
                             ),
                         ],
