+++ lib/modules/settings/routes/configuration_lidarr/pages/connection_details.dart
@@ -25,6 +25,7 @@ class _State extends State<_SettingsConfigurationLidarrRoute> with LunaScrollCon
             key: _scaffoldKey,
             appBar: _appBar(),
             body: _body(),
+            bottomNavigationBar: _bottomActionBar(),
         );
     }
 
@@ -32,6 +33,14 @@ class _State extends State<_SettingsConfigurationLidarrRoute> with LunaScrollCon
         return LunaAppBar(
             title: 'Connection Details',
             scrollControllers: [scrollController],
+        );
+    }
+
+    Widget _bottomActionBar() {
+        return LunaBottomActionBar(
+            actions: [
+                _testConnection(),
+            ],
         );
     }
 
@@ -44,7 +53,6 @@ class _State extends State<_SettingsConfigurationLidarrRoute> with LunaScrollCon
                     _host(),
                     _apiKey(),
                     _customHeaders(),
-                    _testConnection(),
                 ],
             ),
         );
@@ -87,44 +95,41 @@ class _State extends State<_SettingsConfigurationLidarrRoute> with LunaScrollCon
     }
 
     Widget _testConnection() {
-        return LunaButtonContainer(
-            children: [
-                LunaButton.text(
-                    text: 'Test Connection',
-                    onTap: () async {
-                        ProfileHiveObject _profile = Database.currentProfileObject;
-                        if(_profile.lidarrHost == null || _profile.lidarrHost.isEmpty) {
-                            showLunaErrorSnackBar(
-                                context: context,
-                                title: 'Host Required',
-                                message: 'Host is required to connect to Lidarr',
-                            );
-                            return;
-                        }
-                        if(_profile.lidarrKey == null || _profile.lidarrKey.isEmpty) {
-                            showLunaErrorSnackBar(
-                                context: context,
-                                title: 'API Key Required',
-                                message: 'API key is required to connect to Lidarr',
-                            );
-                            return;
-                        }
-                        LidarrAPI.from(Database.currentProfileObject).testConnection()
-                        .then((_) => showLunaSuccessSnackBar(
-                            context: context,
-                            title: 'Connected Successfully',
-                            message: 'Lidarr is ready to use with LunaSea',
-                        )).catchError((error, trace) {
-                            LunaLogger().error('Connection Test Failed', error, trace);
-                            showLunaErrorSnackBar(
-                                context: context,
-                                title: 'Connection Test Failed',
-                                error: error,
-                            );
-                        });
-                    },
-                ),
-            ],
+        return LunaButton.text(
+            text: 'Test Connection',
+            icon: Icons.wifi_tethering_rounded,
+            onTap: () async {
+                ProfileHiveObject _profile = Database.currentProfileObject;
+                if(_profile.lidarrHost == null || _profile.lidarrHost.isEmpty) {
+                    showLunaErrorSnackBar(
+                        context: context,
+                        title: 'Host Required',
+                        message: 'Host is required to connect to Lidarr',
+                    );
+                    return;
+                }
+                if(_profile.lidarrKey == null || _profile.lidarrKey.isEmpty) {
+                    showLunaErrorSnackBar(
+                        context: context,
+                        title: 'API Key Required',
+                        message: 'API key is required to connect to Lidarr',
+                    );
+                    return;
+                }
+                LidarrAPI.from(Database.currentProfileObject).testConnection()
+                .then((_) => showLunaSuccessSnackBar(
+                    context: context,
+                    title: 'Connected Successfully',
+                    message: 'Lidarr is ready to use with LunaSea',
+                )).catchError((error, trace) {
+                    LunaLogger().error('Connection Test Failed', error, trace);
+                    showLunaErrorSnackBar(
+                        context: context,
+                        title: 'Connection Test Failed',
+                        error: error,
+                    );
+                });
+            },
         );
     }
